refactor: minimal knob to [Accordion]
