+++ stories/components/AccordionStory/Basic.tsx
@@ -1,5 +1,5 @@
 import {StoryContainer, StoryTitle} from '../../GlobalStyles';
-import {boolean, number} from '@storybook/addon-knobs';
+import {boolean, number, object} from '@storybook/addon-knobs';
 
 import {Accordion} from '../../../main/components/Accordion';
 import type {AccordionListType} from '../../../main/components/Accordion';
@@ -12,30 +12,17 @@ const ScrollContainer = styled.ScrollView`
   width: 100%;
 `;
 
-const data: AccordionListType = [
-  {
-    title: 'Lists',
-    bodies: ['user', 'mail', 'plan'],
-  },
-  {
-    title: 'mail',
-    bodies: ['mail list', 'category', 'bin'],
-  },
-  {
-    title: 'Reports',
-    bodies: ['report list', 'statistics'],
-  },
-];
-
-const AccordionBasic = (): ReactElement => {
+const AccordionBasic = ({data}: {data: AccordionListType}): ReactElement => {
   return (
     <ScrollContainer>
-      <StoryContainer style={{justifyContent: 'flex-start'}}>
+      <StoryContainer>
         <StoryTitle style={{fontSize: 18, marginBottom: 8}}>Basic</StoryTitle>
         <Accordion
-          data={data}
           shouldAnimate={boolean('shouldAnimate', true)}
           animDuration={number('animDuration', 200)}
+          activeOpacity={number('activeOpacity', 1)}
+          collapseOnStart={boolean('collapseOnStart', true)}
+          data={object('data', data)}
         />
       </StoryContainer>
     </ScrollContainer>
